requirements.txt
boto3
python-nmap
flask